+++ src/components/common/CardSignInRequired.vue
@@ -5,7 +5,7 @@
     </div>
     <div slot="subtitle">
       This page shows data for a specific address. To access the page, please
-      <a @click="goToSession">sign&nbsp;in</a>.
+      <a @click="$router.push(`/welcome`)">sign&nbsp;in</a>.
     </div>
   </TmDataMsg>
 </template>
@@ -14,13 +14,7 @@
 import TmDataMsg from "common/TmDataMsg"
 export default {
   name: `card-sign-in-required`,
-  components: { TmDataMsg },
-  methods: {
-    goToSession() {
-      this.$store.commit(`setSessionModalView`, `welcome`)
-      this.$store.commit(`toggleSessionModal`, true)
-    }
-  }
+  components: { TmDataMsg }
 }
 </script>
 <style scoped>