+++ src/components/common/SessionApprove.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="approve-tran" hide-header>
+  <div class="session-approve">
     <h2>Approve Transaction</h2>
     <div>
       <p>Verify the transaction details below.</p>
@@ -20,9 +20,11 @@
         From
         <Bech32 :address="senderAddress" />
       </div>
-
-      <!-- <TableInvoice :fees="fees" /> -->
-
+      <TableInvoice
+        :amount="amount"
+        :estimated-fee="fees"
+        :bond-denom="bondDenom"
+      />
       <TmFormGroup
         :error="$v.password.$error && $v.password.$invalid"
         class="action-modal-group"
@@ -43,7 +45,7 @@
         />
       </TmFormGroup>
 
-      <div class="approve-tran-footer">
+      <div class="session-approve-footer">
         <TmBtn
           id="reject-btn"
           value="Reject"
@@ -70,34 +72,18 @@ import TmFormGroup from "common/TmFormGroup"
 import TmField from "common/TmField"
 import TmFormMsg from "common/TmFormMsg"
 import LiAnyTransaction from "transactions/LiAnyTransaction"
-// import TableInvoice from "src/ActionModal/components/TableInvoice"
+import TableInvoice from "src/ActionModal/components/TableInvoice"
 import Bech32 from "common/Bech32"
 import { required } from "vuelidate/lib/validators"
-
-// TODO move into own helper file
-// Parse into Lunie tx format from signMessage to display tx properly
-function parseTx(signMessage) {
-  const { msgs, fee, memo } = JSON.parse(signMessage)
-
-  return {
-    tx: {
-      type: "auth/StdTx",
-      value: {
-        msg: msgs,
-        fee,
-        memo
-      }
-    }
-  }
-}
+import { parseTx, parseFee, parseValueObj } from "../../scripts/parsers.js"
 
 export default {
-  name: `session-ext-approve-tran`,
+  name: `session-approve`,
   components: {
     TmBtn,
     TmFormGroup,
     LiAnyTransaction,
-    // TableInvoice,
+    TableInvoice,
     Bech32,
     TmField,
     TmFormMsg
@@ -111,20 +97,23 @@ export default {
     tx() {
       return this.signRequest ? parseTx(this.signRequest.signMessage) : null
     },
-    // fees() {
-    //   return this.tx ? this.tx.tx.value.fee.amount : null
-    // },
+    fees() {
+      return this.tx ? parseFee(this.tx.tx) : null
+    },
     senderAddress() {
       return this.signRequest ? this.signRequest.senderAddress : null
+    },
+    amountCoin() {
+      return this.tx ? parseValueObj(this.tx.tx) : null
+    },
+    amount() {
+      return this.amountCoin ? Number(this.amountCoin.amount) : null
+    },
+    bondDenom() {
+      return this.amountCoin ? this.amountCoin.denom : null
     }
   },
   methods: {
-    setState(value) {
-      this.$emit(`route-change`, value)
-    },
-    close() {
-      this.$emit(`close`)
-    },
     isValidInput(property) {
       this.$v[property].$touch()
 
@@ -136,14 +125,14 @@ export default {
           ...this.signRequest,
           password: this.password
         })
-        this.close()
+        this.$router.push(`/approved`)
       }
     },
     async reject() {
       await this.$store.dispatch("rejectSignRequest", {
         ...this.signRequest
       })
-      this.close()
+      this.$router.push(`/`)
     }
   },
   validations() {
@@ -157,13 +146,13 @@ export default {
 </script>
 
 <style scoped>
-.approve-tran {
+.session-approve {
   padding: 2rem;
   background: var(--fg);
   border-left: 1px solid var(--bc-dim);
 }
 
-.approve-tran h2 {
+.session-approve h2 {
   color: var(--bright);
   font-size: var(--h1);
   font-weight: 500;
@@ -189,7 +178,7 @@ export default {
   flex-direction: column;
 }
 
-.approve-tran-footer {
+.session-approve-footer {
   display: flex;
   justify-content: space-around;
   padding: 1.5rem 0 1rem;