+++ src/components/common/TmSessionHardware.vue
@@ -1,58 +1,49 @@
 <template>
-  <div class="session">
-    <div class="session-header">
-      <a @click="goBack">
-        <i class="material-icons session-back">arrow_back</i>
-      </a>
-      <a @click="close">
-        <i class="material-icons session-close">close</i>
-      </a>
-    </div>
-    <h2 class="session-title">
-      Use my Ledger Nano
-    </h2>
+  <SessionFrame>
+    <div class="session">
+      <h2 class="session-title">
+        Use my Ledger Nano
+      </h2>
 
-    <template v-if="session.browserWithLedgerSupport">
-      <div class="session-main">
-        <HardwareState :loading="status === `connect` ? false : true">
-          <template v-if="status === `connect` || status === `detect`">
-            Please plug in your Ledger&nbsp;Nano&nbsp;S and open the Cosmos app
-          </template>
-          <p v-if="connectionError" class="error-message">
-            {{ connectionError }}
-          </p>
-        </HardwareState>
-      </div>
-      <div class="session-footer">
-        <p class="ledger-install">
-          If you don't have a Ledger Nano, you can
-          <a href="" target="_blank" rel="noopener norefferer">buy one here</a>.
+      <template v-if="session.browserWithLedgerSupport">
+        <div class="session-main">
+          <HardwareState :loading="status === `connect` ? false : true">
+            <template v-if="status === `connect` || status === `detect`">
+              Please plug in your Ledger&nbsp;Nano&nbsp;S and open the Cosmos
+              app
+            </template>
+            <p v-if="connectionError" class="error-message">
+              {{ connectionError }}
+            </p>
+            <TmBtn
+              :value="submitCaption"
+              :disabled="status === `connect` ? false : `disabled`"
+              @click.native="signIn()"
+            />
+          </HardwareState>
+        </div>
+      </template>
+
+      <div v-else class="session-main">
+        <p>
+          Please use Chrome, Opera, or Brave. Ledger is not supported in this
+          browser.
         </p>
-        <TmBtn
-          :value="submitCaption"
-          :disabled="status === `connect` ? false : `disabled`"
-          @click.native="signIn()"
-        />
       </div>
-    </template>
-
-    <div v-else class="session-main">
-      <p>
-        Please use Chrome, Opera, or Brave. Ledger is not supported in this
-        browser.
-      </p>
     </div>
-  </div>
+  </SessionFrame>
 </template>
 
 <script>
 import TmBtn from "common/TmBtn"
 import { mapGetters } from "vuex"
 import HardwareState from "common/TmHardwareState"
+import SessionFrame from "common/SessionFrame"
 export default {
   name: `session-hardware`,
   components: {
     TmBtn,
+    SessionFrame,
     HardwareState
   },
   data: () => ({
@@ -70,21 +61,13 @@ export default {
     }
   },
   methods: {
-    setState(value) {
-      this.$emit(`route-change`, value)
-    },
-    goBack() {
-      this.$emit(`route-change`, "existing")
-    },
-    close() {
-      this.$emit(`close`)
-    },
     async signIn() {
       this.connectionError = null
       this.status = `detect`
       this.address = null
       try {
         this.address = await this.$store.dispatch(`connectLedgerApp`)
+        this.$router.push(`/`)
       } catch ({ message }) {
         this.status = `connect`
         this.connectionError = message