+++ src/components/common/TmSessionWelcome.vue
@@ -1,58 +1,53 @@
 <template>
-  <div id="session-welcome" class="session">
-    <div class="session-header">
-      <a @click="closeSession">
-        <i class="material-icons session-close">close</i>
-      </a>
+  <SessionFrame>
+    <div id="session-welcome" class="session">
+      <router-link to="/">
+        <img
+          class="session-logo-mobile"
+          src="~assets/images/cosmos-wallet-logo.svg"
+        />
+      </router-link>
+
+      <h2 class="session-title">
+        Welcome to Lunie
+      </h2>
+
+      <p class="session-paragraph">
+        The easiest way to send, stake, and vote on the Cosmos Hub.
+      </p>
+
+      <div class="session-list">
+        <LiSession
+          id="creat-new-address"
+          icon="person_add"
+          title="Create a new address"
+          route="create"
+        />
+        <LiSession
+          id="use-an-existing-address"
+          icon="person"
+          title="Use an existing address"
+          route="existing"
+        />
+      </div>
+
+      <p class="footnote">
+        By using Lunie, you accept our
+        <router-link to="/terms" class="link">Terms of Service</router-link> and
+        <router-link to="/privacy" class="link">Privacy Policy</router-link>.
+      </p>
     </div>
-
-    <img class="lunie-logo" src="~assets/images/cosmos-wallet-logo.svg" />
-
-    <h2 class="session-title">
-      Welcome to Lunie
-    </h2>
-
-    <p class="session-paragraph">
-      The easiest way to send, stake, and vote on the Cosmos Hub.
-    </p>
-
-    <div class="session-list">
-      <LiSession
-        id="creat-new-address"
-        icon="person_add"
-        title="Create a new address"
-        @click.native="() => setState('sign-up')"
-      />
-      <LiSession
-        id="use-an-existing-address"
-        icon="person"
-        title="Use an existing address"
-        @click.native="() => setState('existing')"
-      />
-    </div>
-
-    <p class="footnote">
-      By using Lunie, you accept our
-      <router-link to="/terms" class="link">Terms of Service</router-link> and
-      <router-link to="/privacy" class="link">Privacy Policy</router-link>.
-    </p>
-  </div>
+  </SessionFrame>
 </template>
 
 <script>
+import SessionFrame from "common/SessionFrame"
 import LiSession from "common/TmLiSession"
 export default {
   name: `session-welcome`,
   components: {
+    SessionFrame,
     LiSession
-  },
-  methods: {
-    setState(value) {
-      this.$emit(`route-change`, value)
-    },
-    closeSession() {
-      this.$emit(`close`)
-    }
   }
 }
 </script>