+++ test/unit/specs/components/ActionModal/components/TableInvoice.spec.js
@@ -6,25 +6,30 @@ describe(`TableInvoice`, () => {
   const localVue = createLocalVue()
   localVue.directive(`tooltip`, () => {})
 
-  const $store = {
-    getters: {
-      bondDenom: `STAKE`
-    }
-  }
-
   beforeEach(() => {
     wrapper = shallowMount(TableInvoice, {
       localVue,
-      mocks: { $store },
       propsData: {
-        amount: 1,
-        gasEstimate: 1234567,
-        gasPrice: 2.5e-8
+        amount: 17.2,
+        estimatedFee: 0.030864,
+        bondDenom: `STAKE`
       }
     })
   })
 
-  it(`should show the transaction invoice summary`, async () => {
+  it(`should render correctly`, async () => {
     expect(wrapper.vm.$el).toMatchSnapshot()
+  })
+
+  it(`should display the correct subtotal`, async () => {
+    expect(wrapper.text()).toMatch(/17.2[0]+ STAKE/)
+  })
+
+  it(`should display the correct network fee`, async () => {
+    expect(wrapper.text()).toMatch(/0.03086[0-9]+/)
+  })
+
+  it(`should display the correct total`, async () => {
+    expect(wrapper.text()).toMatch(/17.23086[0-9]+ STAKE/)
   })
 })