+++ test/unit/specs/components/common/SessionApprove.spec.js
@@ -9,11 +9,13 @@ describe(`SessionApprove`, () => {
 
   let wrapper, $store
 
+  const signMessage = `{"account_number":"1","chain_id":"testnet","fee":{"amount":[{"amount":"40","denom":"stake"}],"gas":"39953"},"memo":"(Sent via Lunie)","msgs":[{"type":"cosmos-sdk/MsgSend","value":{"amount":[{"amount":"12000000","denom":"stake"}],"from_address":"cosmos1ek9cd8ewgxg9w5xllq9um0uf4aaxaruvcw4v9e","to_address":"cosmos1324vt5j3wzx0xsc32mjhkrvy5gn5ef2hrwcg29"}}],"sequence":"0"}`
+
   beforeEach(() => {
     const getters = {
       signRequest: {
         senderAddress: "cosmos1234",
-        signMessage: "{}"
+        signMessage
       }
     }
 
@@ -25,29 +27,16 @@ describe(`SessionApprove`, () => {
     wrapper = shallowMount(SessionApprove, {
       localVue,
       mocks: {
-        $store
+        $store,
+        $router: {
+          push: jest.fn()
+        }
       }
     })
   })
 
   it(`shows the approval modal with the transaction and an invoice table`, () => {
     expect(wrapper.vm.$el).toMatchSnapshot()
-  })
-
-  it(`should close`, () => {
-    const self = {
-      $emit: jest.fn()
-    }
-    SessionApprove.methods.close.call(self)
-    expect(self.$emit).toHaveBeenCalledWith(`close`)
-  })
-
-  it("moves to other session pages", () => {
-    const self = {
-      $emit: jest.fn()
-    }
-    SessionApprove.methods.setState.call(self, "welcome")
-    expect(self.$emit).toHaveBeenCalledWith("route-change", "welcome")
   })
 
   describe("approve", () => {
@@ -67,9 +56,9 @@ describe(`SessionApprove`, () => {
       await wrapper.vm.$nextTick()
       expect(wrapper.vm.$store.dispatch).toHaveBeenCalledWith(
         "approveSignRequest",
-        { password: "1234", senderAddress: "cosmos1234", signMessage: "{}" }
+        { password: "1234", senderAddress: "cosmos1234", signMessage }
       )
-      expect(wrapper.vm.close).toHaveBeenCalled()
+      expect(wrapper.vm.$router.push).toHaveBeenCalledWith(`/approved`)
     })
   })
 
@@ -79,8 +68,8 @@ describe(`SessionApprove`, () => {
     await wrapper.vm.$nextTick()
     expect(wrapper.vm.$store.dispatch).toHaveBeenCalledWith(
       "rejectSignRequest",
-      { senderAddress: "cosmos1234", signMessage: "{}" }
+      { signMessage, senderAddress: "cosmos1234" }
     )
-    expect(wrapper.vm.close).toHaveBeenCalled()
+    expect(wrapper.vm.$router.push).toHaveBeenCalledWith(`/`)
   })
 })